backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
pydantic==2.9.2
python-dotenv==1.0.1
groq==0.11.0
requests==2.32.3
sentence-transformers==3.0.1
faiss-cpu==1.8.0
orjson==3.10.7
numpy==1.26.4          # (keep 1.x to avoid faiss/numpy 2.x issues)
pillow==10.4.0
pytesseract==0.3.13
pypdf==4.3.1
python-multipart==0.0.9
python-docx==1.1.2
